Rename module "registry" to "registries"
So that modules represent things in aggregate, etc.

Mostly this is because I think it improves the potential usage syntax:

```javascript
const registry = ethpm.registries.get("cool-packages.gnidan.eth");
const release = registry.package("cool-package").release("1.9.1");
```
