+++ test/stub/registries/examples.ts
@@ -1,10 +1,10 @@
 /**
- * @module "test/stub/registry"
+ * @module "test/stub/registries"
  */
 
 import { URL } from "url";
 import * as t from "io-ts";
-import * as registry from "ethpm/registry";
+import * as registries from "ethpm/registries";
 import { StubService as StorageService } from "test/stub/storage";
 
 import StubConnector from "./service";
@@ -12,7 +12,7 @@ import manifests from "test/examples/manifests";
 import packages from "test/examples/packages";
 
 export default class ExamplesConnector extends StubConnector {
-  async connect(options: t.mixed): Promise<registry.Service> {
+  async connect(options: t.mixed): Promise<registries.Service> {
     const storage = new StorageService();
 
     return super.connect({
