+++ test/stub/registries/test/service.ts
